+++ jsonschema-generator/src/test/java/com/github/victools/jsonschema/generator/MethodScopeTest.java
@@ -50,6 +50,8 @@ public class MethodScopeTest extends AbstractTypeAwareTest {
             Arguments.of("getFieldWithPublicGetter", "getFieldWithPrivateGetter", "fieldWithPublicGetter"),
             Arguments.of("isFieldWithPublicBooleanGetter", null, "fieldWithPublicBooleanGetter"),
             Arguments.of("isFieldWithPublicBooleanGetter", "isBehavingSomehow", "fieldWithPublicBooleanGetter"),
+            Arguments.of("isFieldWithMatchingPublicGetter", null, "isFieldWithMatchingPublicGetter"),
+            Arguments.of("isFieldWithMatchingPublicGetter", "isFieldWithPublicBooleanGetter", "isFieldWithMatchingPublicGetter"),
             Arguments.of("getCalculatedValue", null, null),
             Arguments.of("isBehavingSomehow", null, null),
             Arguments.of("isBehavingSomehow", "isFieldWithPublicBooleanGetter", null),
@@ -116,6 +118,7 @@ public class MethodScopeTest extends AbstractTypeAwareTest {
         private long fieldWithPublicGetter;
         @TestAnnotation
         private boolean fieldWithPublicBooleanGetter;
+        private boolean isFieldWithMatchingPublicGetter;
 
         @TestAnnotation
         private int getFieldWithPrivateGetter() {
@@ -128,6 +131,10 @@ public class MethodScopeTest extends AbstractTypeAwareTest {
 
         public boolean isFieldWithPublicBooleanGetter() {
             return this.fieldWithPublicBooleanGetter;
+        }
+
+        public boolean isFieldWithMatchingPublicGetter() {
+            return this.isFieldWithMatchingPublicGetter;
         }
 
         public double getCalculatedValue() {